+++ src/main/java/com/softeer/podo/admin/model/dto/user/LotsUserDto.java
@@ -1,4 +1,4 @@
-package com.softeer.podo.admin.model.dto;
+package com.softeer.podo.admin.model.dto.user;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import java.time.LocalDateTime;
@@ -7,11 +7,10 @@ import lombok.Data;
 
 @Data
 @Builder
-public class ArrivalUserDto {
+public class LotsUserDto {
 	private Long id;
 	private String name;
 	private String phoneNum;
-	private int rank;
 	@JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
 	private LocalDateTime createdAt;
 	private String reward;
